+++ server/src/game/config/commonConfig.ts
