+++ server/src/game/model/bullet.ts
@@ -1,6 +1,6 @@
 import { RectBound } from '../../types/rectBound.type';
-import { CommonConfig } from '../commonConfig';
-import { ServerConfig } from '../serverConfig';
+import { CommonConfig } from '../config/commonConfig';
+import { ServerConfig } from '../config/serverConfig';
 import { OverlapTester } from '../util/overlapTester';
 import { GameObject } from './gameObject';
 import { Tank } from './tank';
