+++ server/src/socket/ioGame.ts
@@ -1,5 +1,4 @@
 import { Namespace } from 'socket.io';
-import { Tank } from '../game/model/tank';
 import RoomManager from '../manager/roomManager';
 
 export default class IoGame {
