+++ server/src/types/room.type.ts
@@ -1,0 +1,11 @@
+import { GameManager } from "../manager/gameManager";
+import { Users } from "./user.type";
+
+export type Room = {
+    roomId: string;
+    users: Users;
+    gameManager: GameManager;
+  };
+  export type RoomMap = {
+    [roomId: string]: Room;
+  };
