+++ server/src/types/user.type.ts
@@ -1,0 +1,8 @@
+export type User = {
+  clientId: number;
+  roomId: string;
+  createdAt: number;
+};
+export type Users = {
+  [clientId: string]: User;
+};
